Use stderr instead of stdout for logging messages
By printing log messages (unconditionally) to the regular output
stream we risk upsetting the program's intended output by adding this
agent. It is more common to print log messages to the stderr stream,
even if they don't indicate any errors. (For example, the
ConsoleHandler in java.util.logging defaults to stderr, not stdout).