Use java.lang.classfile API
Change the bytecode transformation, so that - instead of ASM - it uses
the new java.lang.classfile API. This is faster, results in a smaller
binary, and is hopefully more compatible with future Java versions than
ASM is (as the library gets updated with the runtime).

Note that this requires Java 24, so for the time being, I will create
combined releases that contain both versions.